Start of new design to remove NavigationView - creation of new NavBar
